[fix] fixed scraping video links function and google sheet helper functions
