+++ scrapedVideoLinks.json
@@ -1,0 +1,92 @@
+[
+    "https://www.youtube.com/watch?v=iOmxIV9ehTI",
+    "https://www.youtube.com/watch?v=0DThjRo_oBo",
+    "https://www.youtube.com/watch?v=VyYF9BI3iFA",
+    "https://www.youtube.com/watch?v=tOHr5BJO1co",
+    "https://www.youtube.com/watch?v=6YNYwp57V2Q",
+    "https://www.youtube.com/watch?v=nW6JEbEG7c4",
+    "https://www.youtube.com/watch?v=wFw4TovEicE",
+    "https://www.youtube.com/watch?v=QY_EVSsWL9Y",
+    "https://www.youtube.com/watch?v=YBMvbvTsC_I",
+    "https://www.youtube.com/watch?v=yQd5mWPOKwc",
+    "https://www.youtube.com/watch?v=ULDZk6FoAiY",
+    "https://www.youtube.com/watch?v=HULvf3zLEA8",
+    "https://www.youtube.com/watch?v=p7gtz6JLDuA",
+    "https://www.youtube.com/watch?v=MGMF4oxCBUk",
+    "https://www.youtube.com/watch?v=SzJK24iB-l4",
+    "https://www.youtube.com/watch?v=xD8G_6J3ip8",
+    "https://www.youtube.com/watch?v=0COTZB0iCPU",
+    "https://www.youtube.com/watch?v=gKn9qfNXlhQ",
+    "https://www.youtube.com/watch?v=xn-dM0G2fwA",
+    "https://www.youtube.com/watch?v=c7RPmHF6O1E",
+    "https://www.youtube.com/watch?v=rInqL98Us7o",
+    "https://www.youtube.com/watch?v=QQsSMFHGRUE",
+    "https://www.youtube.com/watch?v=mRVVvX7Wq4c",
+    "https://www.youtube.com/watch?v=_Rq9JXCXGII",
+    "https://www.youtube.com/watch?v=W57Ut45kstM",
+    "https://www.youtube.com/watch?v=Kg9CRnzZjrw",
+    "https://www.youtube.com/watch?v=pQRSLBIPuH8",
+    "https://www.youtube.com/watch?v=p76oc2yfcX0",
+    "https://www.youtube.com/watch?v=n4hv08kNS80",
+    "https://www.youtube.com/watch?v=9nATaCLEVj0",
+    "https://www.youtube.com/watch?v=IJJ5VA1gA34",
+    "https://www.youtube.com/watch?v=Jz8REBaiNUM",
+    "https://www.youtube.com/watch?v=VxVzpKkgOpE",
+    "https://www.youtube.com/watch?v=1MRhcPI6Mw8",
+    "https://www.youtube.com/watch?v=GwagnRjCkXQ",
+    "https://www.youtube.com/watch?v=wNLwOVn0GgM",
+    "https://www.youtube.com/watch?v=FVIDuvA6fpI",
+    "https://www.youtube.com/watch?v=kX2NZq121VQ",
+    "https://www.youtube.com/watch?v=CU_xsEeOLMQ",
+    "https://www.youtube.com/watch?v=wCguPEAavOk",
+    "https://www.youtube.com/watch?v=fLPqmAGWwZI",
+    "https://www.youtube.com/watch?v=c-pazswd3v8",
+    "https://www.youtube.com/watch?v=5exVIYDhhHE",
+    "https://www.youtube.com/watch?v=Pjoew9pcVQU",
+    "https://www.youtube.com/watch?v=4JzMTs7qNzw",
+    "https://www.youtube.com/watch?v=qFjucKqtOX0",
+    "https://www.youtube.com/watch?v=boodj3Zwt6E",
+    "https://www.youtube.com/watch?v=C7bvaUsk7BM",
+    "https://www.youtube.com/watch?v=g3CS8wzmLNk",
+    "https://www.youtube.com/watch?v=uT_igPSTIMk",
+    "https://www.youtube.com/watch?v=1LdYbqLlAlk",
+    "https://www.youtube.com/watch?v=vMW52ZIs3-M",
+    "https://www.youtube.com/watch?v=CsWCCl7C934",
+    "https://www.youtube.com/watch?v=N36dvlNNCaw",
+    "https://www.youtube.com/watch?v=hzP2mM1YZdA",
+    "https://www.youtube.com/watch?v=LPRknsFe96k",
+    "https://www.youtube.com/watch?v=Ct4HkKs_HBY",
+    "https://www.youtube.com/watch?v=eS2IDoafML4",
+    "https://www.youtube.com/watch?v=mba9X3ALPMQ",
+    "https://www.youtube.com/watch?v=4IWdnxvSTkM",
+    "https://www.youtube.com/watch?v=amQm72Nycgw",
+    "https://www.youtube.com/watch?v=CXQwvaKOtNQ",
+    "https://www.youtube.com/watch?v=z783jvfh9FM",
+    "https://www.youtube.com/watch?v=P6Ya0Mj7bAE",
+    "https://www.youtube.com/watch?v=0JZPSlbEiKA",
+    "https://www.youtube.com/watch?v=ToevXdEap2o",
+    "https://www.youtube.com/watch?v=3rZuAl5oNbg",
+    "https://www.youtube.com/watch?v=XjgAGbV3UAQ",
+    "https://www.youtube.com/watch?v=r3i8reFcLx8",
+    "https://www.youtube.com/watch?v=vvgJjEjve9E",
+    "https://www.youtube.com/watch?v=cS991Q0JqTQ",
+    "https://www.youtube.com/watch?v=Fo6wmOyOo0U",
+    "https://www.youtube.com/watch?v=wLAJFTjltbM",
+    "https://www.youtube.com/watch?v=Dps0myF3NNg",
+    "https://www.youtube.com/watch?v=Qm5drcQnVKs",
+    "https://www.youtube.com/watch?v=wwW9IV4V16Y",
+    "https://www.youtube.com/watch?v=CVYzdCvz5UE",
+    "https://www.youtube.com/watch?v=ZBhz9udcREg",
+    "https://www.youtube.com/watch?v=V_wSjgPhbag",
+    "https://www.youtube.com/watch?v=eAZo4jPNRPQ",
+    "https://www.youtube.com/watch?v=2HBwlcpL1tk",
+    "https://www.youtube.com/watch?v=hEgP6YhyR3k",
+    "https://www.youtube.com/watch?v=yBX8U_p2YzU",
+    "https://www.youtube.com/watch?v=EMCQFrp6tHI",
+    "https://www.youtube.com/watch?v=vZXagC-ucEw",
+    "https://www.youtube.com/watch?v=UGvjPyOjLyI",
+    "https://www.youtube.com/watch?v=DfTcn1DMv9g",
+    "https://www.youtube.com/watch?v=mebChjE5Guo",
+    "https://www.youtube.com/watch?v=1ZpYmdtPpWI"
+  ]
+
