+++ src/youtube.helpers.ts
@@ -4,6 +4,8 @@ import * as path from "path";
 import ffmpeg from "fluent-ffmpeg";
 import youtubedl, { Payload } from "youtube-dl-exec";
 
+const CHANNEL_URL = 'https://www.youtube.com/c/Ond%C5%99ejKob%C4%9Brsk%C3%BD'; // Channel videos page  
+
 export async function downloadVideo(
   videoUrl: string
 ): Promise<{ audioPath: string; videoId: string }> {
@@ -148,47 +150,6 @@ export async function downloadVideoInChunks(
     throw error;
   }
 }
-
-// export async function getVideoLinks(channelUrl: string): Promise<string[]> {
-//   try {
-//     const scrapedVideoUrls: string[] = [];
-//     const dateTwoYearsAgo = new Date();
-//     dateTwoYearsAgo.setFullYear(dateTwoYearsAgo.getFullYear() - 2);
-//     const formattedDate = dateTwoYearsAgo.toISOString().split('T')[0].replace(/-/g, '');
-
-//     const videoUrlsString: string | Payload = await youtubedl(channelUrl, {
-//       flatPlaylist: true,
-//       getUrl: true,
-//       noWarnings: true,
-//     });
-
-//     if (typeof videoUrlsString === 'string') {
-//       const videoUrls = videoUrlsString.split('\n').filter(url => url.trim() !== '' && !url.includes('short'));
-
-//       for (const videoUrl of videoUrls) {
-//         const output = await youtubedl(videoUrl, {
-//           dumpSingleJson: true,
-//           flatPlaylist: true,
-//           noCheckCertificates: true,
-//           noWarnings: true,
-//           preferFreeFormats: true,
-//           addHeader: ['referer:youtube.com', 'user-agent:googlebot'],
-//           userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3',
-//         });
-//         if (typeof output === 'object') {
-//           if (output.upload_date > formattedDate) {
-//             scrapedVideoUrls.push(videoUrl);
-//           }
-//         }
-//       }
-//     }
-//     console.log('Scraped video URLs:', scrapedVideoUrls); 
-//     return scrapedVideoUrls;
-//   } catch (error) {
-//     console.error('Error fetching video links:', error);
-//     throw error;
-//   }
-// }
 
 export async function getVideoLinks(channelUrl: string): Promise<string[]> {
   const scrapedVideoUrls: string[] = [];
